nrfx_config_nrf9160: Remove dependency on DT from NRF_* mappings
Stop using base addresses of peripherals extracted from devicetree
in definitions of NRF_* symbols that are needed by HALs and drivers.
Instead, in order to map to secure or non-secure base addresses of
peripherals, take corresponding symbols provided by MDK (with _S or
_NS suffix) depending on the Kconfig option ARM_NONSECURE_FIRMWARE.

This commit provides also temporary mappings of the GPIOTE IRQ number
making it possible to use nrfx_gpiote driver on nRF9160. These mappings
are supposed to be removed when the nrfx_gpiote driver is corrected to
not use the GPIOTE_IRQn symbol.

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>