nrfx_i2s: Introduce NRFX_I2S_STATUS_TRANSFER_STOPPED flag
Add the NRFX_I2S_STATUS_TRANSFER_STOPPED flag to the nrfx_i2s driver's
API, to provide the application with a notification when the driver
actually stops the transfer after the nrfx_i2s_stop function is called.

Move also the corresponding update of the driver's state to the point
before the handler function is called, so that it is possible to start
a next transfer directly from the handler function.

This is a temporary patch, needed to proceed with a PR in the main
Zephyr repository that requires such extension of the nrfx_i2s driver.
Same changes will be contained in the next release of nrfx.

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>