nrf_radio_802154: Replace CONFIG_ENTROPY_NAME with DT_LABEL(<rng_node>)
The CONFIG_ENTROPY_NAME symbol is no longer used in the Zephyr tree.
Replace it with a direct reference to the label property of the only
instance of "nordic,nrf-rng" compatible node in dts, as there is no
advantage in using the generic symbol that replaced CONFIG_ENTROPY_NAME
(i.e. DT_ENTROPY_NAME) in an nRF specific driver.

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>