nrfx_glue: Fix NRFX_PPI_GROUPS_USED_BY_BT_CTLR definition
This is a follow-up to commit c84263d8ecc7767b25d9abab9502c9e748d045e9.

The same correction that was done for the PPI channels macro needs to
be done also for the PPI groups macro.
This commit reorganizes a bit definitions of these macros in question,
for easier maintenance.

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>