nrf_gpio: Prevent warnings about switch cases falling through
Add special comments that will prevent warnings that would otherwise
appear when compilation is done with the -Wimplicit-fallthrough
options, as the falls through are intentional.

This is a temporary change in a file imported from the nrfx repository
and it is supposed to be overwritten by the next update of nrfx.

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>