drivers: nrf_radio_802154: Ensure RAAL finishes its processing
When the nRF 802.15.4 radio driver is used in multiprotocol scenarios,
it might happen that the arbiter revokes access to the RADIO peripheral
and it is not notified to the higher layer, e.g. because of other preconditions
being denied already. This patch enables the arbiter to complete its processing
even if the higher layer is not notified about it.

Signed-off-by: Jedrzej Ciupis <jedrzej.ciupis@nordicsemi.no>