nrfx_glue: Exclude DPPI resources used by BT_CTLR from allocation
Use for DPPI the same mechanism that was so far used for PPI to prevent
assigning of channels and groups used by the Bluetooth controller to
other requesters.
Remove also the additional condition (CONFIG_BT_LL_SW_SPLIT) so that
this mechanism can be used for other Bluetooth controllers as well.

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>