drivers: nrfx_{twi, twim}: Fix busy flag not reset
This is necessary so that next occurence of nrfx_{twi, twim}_enable()
followed by {twi, twim}_xfer() will not raised a NRFX_ERROR_BUSY error.

This could happen when a previous transfer is started but is aborted
using nrfx_{twi, twim}_disable() before the end of the transaction which
calls {twi, twim}_irq_handler().

Signed-off-by: Xavier Chapron <xavier.chapron@stimio.fr>