drivers: Fix several radio driver related bugs
Fix #1:
Transmission failure would cause to leave an allocated buffer resulting
in a serialization failure leading to an assert, which was seen as a
timeout error on APP core side.

Fix #2:
It is possible for nrf_802154_spinel_send function to be preempted
resulting in the static frame buffer being reused. The buffer size was
reduced accordingly to 802.15.4 frame size and possible additional data.

Fix #3:
Radio driver's API call using serialization can be preempted after
issuing a serialized call request to NET core, but before preparing
the notifier for an awaited property. This results in receiving a
property that no one waits for and dropping the property as an unwanted.
In current implementation the awaited property is prepared during the
notifier lock.

Re-fix #4:
The issue which manifests itself in disabled NVIC interrupts in SED
roles.

Signed-off-by: Pawel Kwiek <pawel.kwiek@nordicsemi.no>