drivers: nrf_radio_802154: Enable 802.15.4 resources validation
Provides bit masks which are intended for use in the nrfx_glue.h
to mark the (D)PPI channels used by the 802.15.4 driver as occupied
and thus unavailable for allocation through nrfx_ppi.

Verification if resources used by the 802.15.4 driver do not overlap
with those used by the mpsl (if mpsl enabled).

Signed-off-by: Adam Zelik <adam.zelik@nordicsemi.no>