twis: Adding user-defined context to be used in callbacks
The callback called each time there is an eve,t in the twis
module is not context-aware as it is impossible to give
it any contextual data.

When this module is used with Zephyr (even if not officially
supported), this is an issue as it is impossible to get
back the pointer to the device structure corresponding to
the device which fired the event.

If only 1 twis module is used, it is not an issue as we
can define a static variable that can be directly called
in the callback.

However, if multiple driver instances are defined, it
becomes impossible to know which instance fired the
callback.

The changes are meant to be backward compatible as it
does not change any function of typedef definition

Signed-off-by: Giuliano Franchetto <giuliano.franchetto@intellinium.com>