nrfx_spim: Fix handling of anomaly 109 on nRF52
The special part of `irq_handler` that handles auxiliary zero-length
transfers has to be executed only when interrupts on the STARTED event
are enabled. Otherwise, each transfer will be repeated indefinitely.

And the STARTED event needs to be cleared before related interrupts
are enabled, as for normal transfers this event is not cleared in the
interrupt handler.

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>