nrfx: Fix overflow of pin_flags array on nRF52820 in GPIOTE driver
GPIOTE driver was incorrectly defining the size of the `pin_flags` array
on nRF52820 SoC. The target has 18 pins - as the value of
the `P0_PIN_NUM` symbol. However, those pins are not consecutive pins
from 0 to 17. The nRF52820 has routed pins 0-8, 14-18, 20 and 28-30.

As a result, accessing pin 18 and above marks in reads/writes
that exceed the boundary of the `pin_flags` array.
To mitigate that issue, the `MAX_PIN_NUMBER` symbol that is used
to define the `pin_flags` array size is now defined as 32 for
the nRF52820 device.

Signed-off-by: Adam Wojasinski <adam.wojasinski@nordicsemi.no>
Signed-off-by: Nikodem Kastelik <nikodem.kastelik@nordicsemi.no>