nrfx_uarte: Fix race condition and optimize isr
If interrupt was executed exactly when ENDRX occurred it is possible

that RXSTARTED (which is read before ENDRX) is read as false but it
actually occurred (if there is a linked reception). It must be read
again to ensure that it is handled in the same interrupt call.
Otherwise there is a risk that on next interrupt it will be too late
to handle it.

Additionally, Read int mask only once in the irq handler to reduce
time spent in the handler.

Signed-off-by: Krzysztof Chruściński <krzysztof.chruscinski@nordicsemi.no>