nrfx: hal: nrf_comp: patch main_mode and speed_mode set
The functions for setting the main and speed modes don't clear their
fields before setting the desired value. This results in not being
able to set the main mode to single-ended (0) after having set the
mode to differential (1), similarly, the speed mode can not be set
to low (0) after being set to medium (1) or high (2), and it can be
set to the reserved value 3.

Signed-off-by: Bjarki Arge Andreasen <bjarki.andreasen@nordicsemi.no>