nrfx: hal: nrf_grtc: Remove readback from event clear
Readback was added to ensure that interrupt is not triggered spuriously.
It may only happen if event clearing is done just before exiting the
interrupt. It is not the case in the GRTC driver. Each access to the
register interface is costly (~160 320MHz cycles) so it must be avoided
if possible.

Signed-off-by: Krzysztof Chruściński <krzysztof.chruscinski@nordicsemi.no>