nrfx: mdk: fix cmsis include for BSIM/POSIX
What is the change?
Do not use header files from cmsis module when compiling for BSIM,
instead use zephyr/boards/native/nrf_bsim/common/cmsis/cmsis.h
which has the required alternatives apis.

Why do we need this change?
BSIM ci reported compiler errors when CMSIS_6 module was used for
Cortex-M instead of the older cmsis module.
The error was because CMSIS_6 couldn't find "arm_acle.h" and that
__ARM_ARCH_PROFILE was not defined to an expected value however,
this error highlighted that hal_nordic was including cmsis header
files even CONFIG_ARCH was "posix" and not "arm".

Updating to CMSIS_6 is important for future maintenance and to add new
features so we can't include those header files in hal_nordic but,
hal_nordic still needs some cmsis alternative header file to compile.
Upon inspection it was found that there is a cmsis.h created as an
alternative when compiling for POSIX/BSIM in Zephyr so,
we use that file and fix the issue.

Signed-off-by: Sudan Landge <sudan.landge@arm.com>