nrfx: gppi: Improve nrfx_gppi_domain_channel_get
nrfx_gppi_domain_channel_get could be used to get exact DPPI channels
that were used in the connection handle. Connection handle was also
using PPIB channels and those could not be extracted from the connection
handle. This patch extends functionality of nrfx_gppi_domain_channel_get
to allow also getting PPIB channel used in the connection.
Name of the function is not changed as this additional functionality
is rare to be used so only description is extended.

Signed-off-by: Krzysztof Chruściński <krzysztof.chruscinski@nordicsemi.no>