nrfx: gppi: Fix use of nrfx_gppi_ext_conn_alloc
When external resource in one of the domains was given then it was
expected that no resource will be touched in that domain but gppi
was touching (enabling, disabling and freeing) channel 0 in that domain.
Fix by using a reserved value to indicated that channel in that domain
is reserved outside.

Signed-off-by: Krzysztof Chruściński <krzysztof.chruscinski@nordicsemi.no>